commit7:update and delete sellers feature added

+++ Backend/src/main/java/com/fcom/repository/AdminRepository.java
@@ -6,8 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.fcom.model.admin;
+
 @Repository
-public interface AdminRepository extends JpaRepository<admin, Integer> {
+public interface AdminRepository extends JpaRepository<admin, Long> {
 	Optional<admin> findById(Long id);
 
 }
